requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
datasets==2.13.1
dill==0.3.6
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.2.0
huggingface-hub==0.15.1
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.3.1
jsonlines==3.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.6.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.8.1
numpy==1.19.5
openai==0.27.8
packaging==23.1
pandas==1.4.4
Pillow==10.0.0
pyarrow==12.0.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rank-bm25==0.2.2
regex==2023.6.3
requests==2.31.0
rouge-score==0.1.2
sacremoses==0.0.53
scikit-learn==1.3.0
scipy==1.5.4
simcse==0.4
six==1.16.0
smart-open==6.3.0
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==2.0.1
torchdata==0.6.1
torchtext==0.15.2
tqdm==4.65.0
transformers==4.10.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
